backend/requirements.txt
gunicorn
PyMuPDF
Pillow
pandas
pdf2docx
Flask
Flask-Cors
Werkzeug
markitdown
